+++ src/scripts/api_scripts.h
@@ -13,6 +13,7 @@
 
 #include <stdbool.h>
 
+bool scripts_file_reload(struct t_scripts_state *scripts_state);
 bool scripts_file_read(struct t_scripts_state *scripts_state);
 bool script_save(struct t_scripts_state *scripts_state, sds scriptname, sds oldscript,
         sds file, int order, int version, sds content, struct t_list *arguments, sds *error);