+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This is a small bug fix release.
 
 ### Changelog
 
+- Feat: New API method MYMPD_API_SCRIPT_RELOAD to reload scripts from disk #1327
 - Upd: zh-Hans translation
 - Fix: Implicit conversion creating keys on systems with OpenSSL 1.1.x #1323
 - Fix: Request stored playlist entries in chunks (MPD 0.24)